Kotlin: Extract LighterAST comments as well as PSI comments
