+++ java/kotlin-extractor/src/main/kotlin/utils/versions/v_1_4_32/CommentExtractorLighterAST.kt
@@ -1,0 +1,11 @@
+package com.github.codeql.comments
+
+import com.github.codeql.*
+import org.jetbrains.kotlin.ir.declarations.*
+
+class CommentExtractorLighterAST(fileExtractor: KotlinFileExtractor, file: IrFile, fileLabel: Label<out DbFile>): CommentExtractor(fileExtractor, file, fileLabel) {
+    // We don't support LighterAST with old Kotlin versions
+    fun extract(): Boolean {
+        return false
+    }
+}
